+++ checklist-app/src/main/java/com/datapath/checklistapp/dao/repository/ChecklistAnswerRepository.java
@@ -3,5 +3,5 @@ package com.datapath.checklistapp.dao.repository;
 import com.datapath.checklistapp.dao.entity.AnswerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChecklistAnswerRepository extends JpaRepository<AnswerEntity, Long> {
+public interface ChecklistAnswerRepository extends JpaRepository<AnswerEntity, Integer> {
 }
