+++ checklist-app/src/main/java/com/datapath/checklistapp/dao/repository/QuestionExecutionRepository.java
@@ -3,5 +3,5 @@ package com.datapath.checklistapp.dao.repository;
 import com.datapath.checklistapp.dao.entity.QuestionExecutionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionExecutionRepository extends JpaRepository<QuestionExecutionEntity, Long> {
+public interface QuestionExecutionRepository extends JpaRepository<QuestionExecutionEntity, Integer> {
 }
