+++ checklist-app/src/main/java/com/datapath/checklistapp/dao/repository/QuestionSourceRepository.java
@@ -3,7 +3,7 @@ package com.datapath.checklistapp.dao.repository;
 import com.datapath.checklistapp.dao.entity.QuestionSourceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionSourceRepository extends JpaRepository<QuestionSourceEntity, Long> {
+public interface QuestionSourceRepository extends JpaRepository<QuestionSourceEntity, Integer> {
 
-    QuestionSourceEntity findByIdentifier(Long identifier);
+    QuestionSourceEntity findByIdentifier(Integer identifier);
 }
